remove the cookie instead of resetting its value.
prior to that, it caused a weird behavior. assume you are authenticated (having a valid not expired QuarkusUser cookie), and then you rotate the public/private keys (even by restarting quarkus in dev mode) and then you visit a page that requires authentication, so it tries to authenticate and since the QuarkusUser value was created with the old keys, the cookie becomes invalid, so the method AuthenticationFailedExceptionMapper.authenticationFailed was invoked, which, in turn called AuthenticationFailedExceptionMapper.redirectToRoot method setting the QuarkusUser to an empty string "" and also correctly redirecting to "/".

however, now, QuarkusUser has an empty value, and if you try to access a protected page with it, it triggers authentication checks and then redirecting to "/". on the contrary, it should redirect to the login page if the @LoginPage existed or to a 401 Unauthorized error. (as if the QuarkusUser cookie did not exist)

so the solution is to remove the cookie (instead of resetting its value to an empty string), so the authentication won't kick in and then properly redirecting to the login page or return 401 Unauthorized error.